Work solely with reported identifier
i.e. Work solely with the primary ID reported by identifier-binding,
not the nominal-id.

To do so, we need to look for `make-provide/contract-transformer` in
the fully-expanded syntax, and hope that it always adds the weird ".1"
suffix that it seems to do in my observations. Obviously this is a
kludge that needs more investigation.

This commit does break the `from-m` test case, so that needs to be
investigated more.

Also, for now this commit does not remove the extra `sourcesym` column
in the `uses` table. I think that can be done, but I want to solve the
`from-m` issue first.