Various changes
Move our hack to add definitions for contract wrapper functions to its
own file --- and don't use it when we have a new-enough
drracket/check-syntax.

Add a couple smoke tests, for get-uses and get-uses/transitive.

Fix a match pattern bug in use-pos->def, and rewrite it to replace the
named-let with a recursive call to itself.

Performance:

Replace a db query with a hash-table.

Set sqlite "pragma synchronous = off". This makes a huge difference
given how many inserts we perform.