Handle sub-range-binders, for defs but not names
This is sort of an interim, checkpoint commit:

- There are duplicate flavors of some tables/views, e.g. both
`def_arrows` and `sub_range_def_arrows`. The former aren't really
used, and I should be brave enough to remove them.

- So far this is only for the definition graph -- not the name
introduction graph. I haven't yet tried to do that. That's a big
omission because, although one of the motivations for handling
sub-range-binders is for more-precise visit-definition using the
definition graph, the other motivation is for reliable multi-file
rename commands.