Re-organize not to use background thread to analyze more files
Eliminate the "analyze more files" thread.

Instead, when analyze-path discovers more files to analyze, it adds
them to the `digests` db table (if not already there) using a dummy
digest that will never match the actual file digest.

Then, when analyze-all-known-paths is called, it will handle them.
Furthermore, it recurs -- in case any of those analyses discovered
more files -- until eventually it arrives at the fixed point where
no more analyses occurred. (As before, analyze-path returns a boolean,
did it analyze something or not.)

This way, if the user wants to analyze more files in the background,
it is in control of how/when that happens. It can use a thread it
manages, for example.

This also is nice because when testing or developing, it's clear when
the "analyze all files" process does conclude -- it's simply when
analyze-all-known-paths returns. At that point you can explore, run
certain tests, etc.