Better attempt at handling phase levels
Change analyze-more to use more of the "outer structure" of the
analysis done in drracket's traversals.rkt. For one thing, this is
necessary to handle phase levels properly (accounting for things like
begin-for-syntax and module forms). Also, we may as well start
preparing for the possibility that someday this would become a PR to
change traversals.rkt.

As of this commit, phase levels are stored for both definitions and
exports. However only defs_xrefs and supporting views have been
updated to do phase-aware queries. Still to-do: Updating name_xrefs. I
haven't fully thought through how that will work, wrt to require and
provide forms like for-{meta template syntax}. Whereas we store and
use identifier-binding's `from-phase` for definitions, we store (and
presumably need to use, somehow TBD) both the `nominal-import-phase`
and `nominal-export-phase` values for exports and the name graph. This
may turn out to be straightforward, I just haven't tackled it yet.