Actually use spaces
The previous commit simply updated our #%require and #%provide form to
accept the new space sub-forms, but didn't attempt to use the spaces
or the racket/phase+space functions provided to manipulate
phase+spaces.

This commit does attempt that, but:

- I'm not confident it's correct.

- I'm squishy on the distinction between phase+space+ and
phase+space-shift+.

- Although I added a test, I don't yet have a good example exercising
our handling of spaces. For phases I have a test where the same
identifier is defined and used for runtime and for-syntax. Although I
understand how to provide and/or require something for-space, I don't
yet understand the equivalent of `begin-for-syntax` (which is for
phases) for spaces.