Delete for-space tests
If I recall correctly, the test example was based on documentation
from a pre-release version of Racket. The example itself is an error
in current, released Racket (e.g. 8.7).

For now I'm just deleting this, until I have time to refresh my memory
about "spaces" and create a valid up-to-date example to test.