Remove support for older versions of racket and drracket
In this commit I'm ripping out some attempts to support older versions
of Racket. Why: We will effectively need a very new version of
drracket-tool-text-lib for all "arrows" to work properly, including
for renaming imports and exports. I'd like to focus on getting things
working properly for the latest versions, and meanwhile simplify the
code. If support for older versions turns out to be important, after
all, I can focus on that, then, and use some of the commit history as
a head start.