Only return strings from `str`
I had some code to return numbers as numbers; IIRC I added this when
enhancing phase to be phase+space. But now I think it's fine for phase
to always be a string, even when it's just a number (for the #f space
case). It's on my to-do list to review the whole business with spaces,
anyway.