Various performance improvements
- Don't use a proxy place or thread for sqlite.

- Do nest all the inserts in a transaction, to avoid per-insert commit
  overhead.

Despite this, analyzing a file like this db.rkt can insert about 2,000
arrows (1,000 each def and name), and this seems to take about 1.5
seconds. This isn't delightful. Although the subsequent query perf
seems great, writes can happen many, many times e.g. while a file is
being edited, and that perf matters, too.

[I wonder if I should explore _not_ using sqlite, and instead using
in-memory data structures that are serialized between runs. The
on-disk locs.sqlite is about 40 MB for this project plus transitive
analyses into Racket itself. Maybe it's not crazy to think of doing
this all in-memory (?).]