No longer use sqlite db
This commit explores using data structures in memory, rather than a
sqlite db.

As of this commit two tests fail, related to re-provides. The old
design used arrows with fake, negative positions. I haven't yet
decided on a better approach. I want to get an interim commit as a
baseline before working on that.

This commit has a number of limitations:

- No (de)serialization. Obviously the time/space to do so would be a
significant consideration.

- Things are optimized solely for the use -> def access pattern. That
is, there are interval-maps with [use-beg use-end) indexes, and the
[def-beg def-end) positions are in the value in the interval-map;
searching for defs by position requires iterating the whole
interval-map dictionary.

- There is no particular effort to index exported bindings; to find
all files that use one, we must iterate all files' data structures.

- Some of the arrow structs could be rationalized better. For example
a lexical-arrow cannot, by definition, have differing use and def
symbols, so it is a waste to store both. Also we probably don't need
the "full" use and def stuff stored, as it is either unused or
redundant. (For instance when sub-range-binders are involved, we store
"pre:foo" as the full text, whereas there are actually two arrows,
each with "pre" and "foo" as their symbol.)