Fix problems from commit 5f0b0eb storing imports as symbols
Also, note that `imports` is intended to be used simply for
completion candidates.

I had a fuzzy idea that it could also be used as a hint as to which
files use a certain import, and used the `ibk` struct to store also
the submods and phase+key. However those latter are about the import
site, not the export site as from identifier-binding. So in this
commit we simplify `imports` to be just a set of symbols -- completion
candidates -- like it was before in the sqlite approach (though
strings not symbols in that case).

As for my fuzzy idea, how to optimize which files to check: Whenever
we create an import-arrow from check-syntax, we could also update some
additional "index".  For example we could have a new `used-imports`
field, which is a (setof ibk?), as a quicker test prior to digging
through all the arrows looking for import arrows. Or similar ideas.