Require newer Racket and drracket-tool-text-lib
This continues what some earlier commits already started: Dropping
some attempts to be compatible with older versions of Racket. Although
normally I'm a huge fan of that with my packages, including using CI
runs to test against older versions, in this case it's not really
practical. We need too much, and some of the changes are in Racket
itself (such as a change to allow arrows to point to contract
wrappers).

Note that the info.rkt is not really tested. This is nowhere near
being ready for real use by other people. Mainly I wanted to document
the required version numbers.