Add tests for prefix-in and prefix-out
Unfortunately the tests for prefix-out cannot pass unless it is
changed, as well as all-defined-out. See comment.

[Although this seems like this could be a simple change in
reqprov.rkt, just adding sub-range-binders property, I think it's
complicated by the "provide transformer" system and make-export's
treatment of things as symbols not syntax. So would need to check with
Matthew and/or Robby.

The all-defined-out part might be simpler -- maybe just giving correct
srcloc to the identifiers in the #%provide rename clauses -- but I
haven't investigated carefully.]