errors: Record both source path and error path
To save space, when they're the same, we record #f for the error path.

This means that if a file has an error, and N files import it, the
error will be recorded for each of the N files. Although maybe that
could be space-optimized, conceptually it's what we want: When e.g. a
user is editing a file, we want to be able to show them any/all
errors, including in imported files.