New design backed by sqlite db instead of rktd.gz file
Use a very simple sqlite table where each row corresponds to an
analyzed file, with two columns, "path" and "data". The latter is the
`file` struct, serialized, written, and gzipped.

So effectively this is like the previous commit, except we can
load/store individual file data. The `files` hash-table acts more
like a cache for a small working set of these. It no longer always
gets all known files (which for main-distribution plus a few installed
packages can be on the order of 8,000 files).

The idea is that there are two main access patterns. In one, the user
has a relatively small working set of files they are editing and
browsing. That set can fit in the cache. At the same time, the
rename-sites command may need to search very many files for uses of a
definition; it can do so without disturbing the small working set
cache, instead getting the file data one by one from the db.

At least that's the theory. Still exploring.