Re-introduce index of nominal imports
This records which files nominally import an export. It can be used by
rename-sites' helper function def->uses/same-name to narrow down which
files' analyses need to be examined to find uses.

When I had tried this before, it didn't seem to help rename-sites work
much faster. But it definitely helps with say 8,000 files in the db.

Before, I had stored this as an extra hash-table within the main file
struct. This time, I'm storing it in a separate sqlite db, using
normal relational tables. Not only is the storage separate, the code
is structured so that analyze-path gather this data, then submits it
all at once for storage. That way, someday we could instead defer this
to be done lazily, to avoid delaying returning the main analysis
results to the client. In other words, when a user is editing a file,
they want the main results available as soon as possible. The nominal
imports index is only used by rename-sites, and only needed if/when
that is ever run. TL;DR this index used only by rename-sites need not
slow down all other functionality.

(I can also imagine an option where we can be configured not to store
this, at all, say for clients that don't care about using rename-sites
at all.)

So far the extra time doesn't seem too bad, so for now we do store it
eagerly. Typically not too many rows need be inserted. Plus the batch
approach allows inserting all rows as part of a single transaction,
which IIUC helps with sqlite insert speed.