Rename some things
Use a prefix style to distinguish similar functions from different
modules (as opposed to stuff like * suffix).

e.g. Rename get-file* in store.rkt itself to get-file, and, prefix it
as store:get-file when imported into other modules that have their
own, higher level get-file function.

e.g. When two imported modules both define an "open" function, prefix
each when importing, and use in the definition of a "higher level"
"open".

(I'm often a fan of this because that way each "component" module can
be written independently using very simple names. Then importers can
be responsible for prefix-renaming as/when needed to avoid collisions,
and/or when desired to make their code clearer.)