Change how arrows are stored
Before we used an interval-map keyed on the use site interval.

Drawbacks:

- When we needed to look up things by def site, we had to iterate the
whole thing.

- An interval-map lacks the ability to look up, then continue
iterating from that position. That's why we've already started
changing some other things to use our "span-map".

So now store arrows using a pair of span-maps, one for def->uses and
another for use->def.

At the same time, store both the use and def spans in the `arrow`
struct. Yes it's redundant given those values being stored as interval
keys in each span-map, but no I don't think the space really matters.

Speaking of space, we store the same immutable arrow struct in both
maps; IIUC this is just the same pointer in both. (But even if it were
a deep copy, I don't think it matters. Our new approach is to cache a
small number of these file structs; most live only in the db. And for
serializing, all we need is one list of arrows; the dual maps can be
constructed from that one list upon deserialization.)