Improve concurrency and break-ability
The only sensitive part is when we update the stores.

What I'm thinking ahead to, here, is that Racket Mode's back end runs
commands on multiple threads. Furthermore when the front end requests
a check-syntax for a path, but one is already underway for that path,
it break-threads the latter to abandon it (no point finishing for an
outdated version). I think this commit should suffice to support that.