Add get-point-info, as used by new racket-pdb-mode
The new mode is on the `pdb` branch of the racket-mode repo. (Currently
that's commit 86e35ea, although that's likely to change if we rebase
that branch on the main branch.)

The old get-annotations is still available in case that's something
other clients might want. Theoretically it could even be used by a
client to do something like get-point-info, just less efficiently.

Again, all this is still WIP, even though it's getting somewhat closer
to being able to survive in the real world.