Work around problem with syncheck:find-source-object
I'm not sure this needs to be the final answer. If Robby is
willing/able to adjust this in drracket/check-syntax, and bump its
version, we could use that. (We're already requiring a very new
version of drracket/check-syntax, as well as a fairly new version of
Racket itself.)

Also delete the helper function string->syntax, folding it into
analyze-code. (Maybe that should be a separate commit. It's related
only in that I was considering all sorts of theories about the
problem, including whether certain parameterizations weren't in
effect. I decided it was easier to reason about this without detouring
to the separate helper function.)