Refactor db connection; add plumber
A previous commit had eliminated the explicit `open` function but
`close` remained. Instead handle disconnect via plumber.

Also refactor the maybe-create/connect pattern to a helper function
used by both store.rkt and nominal-imports.rkt.

(As was already the case prior to this commit, each is a separate db.
The rationale is a scenario where some client doesn't care about
rename-sites, which is the only use of the latter. With a little work
we could support configuring the latter not to be created/updated.
Which wouldn't save a ton of time or space, but someone might care?)