Interim commit; record more information
This is a step toward the overal theme of capturing "all" data given
to syncheck methods, so we can "play it back" to clients like DrRacket
that want to keep using that "classic" interface. As a result we must
store more in the db than before -- more than we need for rename-sites
analysis or for clients lke Racket Mode for Emacs.

Capture all of the syncheck:add-arrow/pxpy/name-dup and
syncheck:add-doc-link values (except name-dup which is not
serializable).

We need to handle more methods (except unused and/or unserializable
ones like add-id-set), but I want to commit this before detouring to
handle zero-width items. (I had started to attempt that, but it wasn't
working properly. I was juggling a mess of too many changes, and
things like rename-sites was broken by the presence of zero-width
arrows which aren't even relevant for renaming :(. I manged to back
out enough to be able to make this commit.)

Start to flesh out a test, which won't pass yet due to zero-width
items, as well as needing to flesh out more syncheck methods.