read-file-from-sqlite: Add exn handler
When there is an exn deserializing, log a warning and return false --
i.e. act as if it were not in the db. That should cause a new analysis
to occur, and later be written, replacing the outdated one.

This suffices for changes to the `file` struct or its components that
cause an exn, such as adding a field to the `file` struct. But less
blatant changes remain a potential problem. So I need to do a fuller
think about "database migration".

The good news is that "migration" per se isn't strictly necessary. In
principle the db is just a cache; everything can be recreated assuming
the original .rkt source files still exist (and if they no longer
exist, it is actually a mis-feature for the db to still refer to
them). Of course, if someone is maintaining a huge db of say 10,000
files that take hours to re-analyze, migration per se would be, if not
strictly a must-have, then an extremely-nice-to-have time-saver.