Raw capture syncheck arrow, jump, and prefixed require reference
Because we adjust what synheck:add-arrow reports -- we add or move
some arrows for renaming imports or exports -- we can't later use our
own arrows to synthesize the original syncheck:add-arrow for a client
we wants that interface. Likewise if we synthesize syncheck:add-jump
from our own arrows. Also, we haven't yet tried to support
syncheck:add-prefixed-require-reference, because we don't use it
ourselves.

So add three new fields to the `file` struct, in which we just capture
original values. We don't even bother to adjust the beg/end coords
from 0-based to 1-based, like we usually do, because all we're going
to use this for is to "play it back" to a syncheck:annotations object.

This lets us pass the remaining test items for our "classic" syncheck
style API.