Move db files
These had been stored in a `data` subdir of the source code, which
obviously isn't practical for real world use.

Instead:

Use a PDB_DIR environment variable, if any, for the directory to
hold our data files. Otherwise...

Use (find-system-path 'cache-dir) if it exists, else 'home-dir, as the
parent directory for a "pdb" subdirectory to hold our data files. So
on *nix for example the parent might be ~/.racket/, and our files will
go in ~/.racket/pdb/.

In any case, the directory is created if it doesn't exist.