Change overall control flow
Previously:

analyze-path would return a boolean whether it analyzed anything, and,
handle other discovered files (direct imports of the file being
analyzed) by writing "dummy" `file` instances to the cache/db. The
idea being the client would need to all an analyze-all-paths function
to analyze those eagerly, if desired.

Now:

1. analyze-path starts a thread for each analysis. When the
direct-imports? argument is true, it goes ahead and starts analysis of
those imports (otherwise it does nothing). So if a user opens a file
in a client editor, it can proactively analyze imports on the theory
that one of those files might be one the user will open next. And in
any case knowledge of files involved in the user's project will grow,
even if they don't yet add-directory some entire bunch of files.

2. add-directory now means calling analyze-path for each file.

3. Similar to how Racket Mode's back end already does this, when
analyze-path is called for a path that is already being analyzed, the
older thread is canceled. So when user is editing the file, there is
no point in continuing the outdated analysis; just abandon it and let
the new one proceed.

---

Probably a future commit can get closer to not exposing analyze-path
at all. Instead, can just expose things like the query API, syncheck
API, use->def, and rename-sites. These already call `get-file` in
analyze.rkt, which either returns the already-analyzed file from cache
or db, else run analyze-path.

A (somewhat) related change: We could adapt the syncheck API so that
when the file isn't read from cache/db, the initial analysis "play
back directly" to the client's annotations object. This would reduce
the latency to be much closer to what it is when a client uses
drracket/check-syntax directly. (Whereas status quo we analyze, create
the file object, and only then synthesize the method calls.)