Move digest out of `file` struct to its own db column
Motivation: When something like analyze-path wants the file data only
when it matches a specific digest, in the case where it does not match
we needn't read/unzip/deserialize/massage data that won't even be
used.

Also: Add a simple versioning mechanism to detect when the db is no
longer valid and must be re-created.