analyze-path: Provide expanded syntax when fresh analysis
Based on discussion with Robby:

  https://racket.discourse.group/t/syncheck-find-source-object/1829/39

Now analyze-path returns

  (or/c fresh-analysis?
        cached-analysis?
        exn:break?)

When the value is fresh-analysis?, it may be given to
fresh-analysis-expanded-syntax to get the fully expanded syntax.

Whereas when the analysis was obtained from cache/db, no expanded
syntax is available.

(Both fresh-analysis and cached-analysis have a file? struct value,
but that is an internal value not exposed in the public API.)