sqlite connection: Use place not os-thread
Otherwise, in the following commit, calling (collect-garbage 'major)
causes an error about "no engine started".

Of course I want to investigate that, report the problem, and (if
possible) use an os-thread instead of a place. But at the moment I
have too many balls in the air so I'm just working around this.