raco setup: Don't use --avoid-main
I cargo culted this from previous Makefiles; I don't recall don't who
advised me to use this, and why. But it seems to be preventing
compiling -- which I don't remember happening before.

Anyway this caused me a lot of confusion when I was trying to
understand why pdb was slow to start up, even after I (supposedly)
compiled ahead of time.

For now I'm just removing this, because I don't want to get
side-tracked. Anyway this is in the Makefile used for development and
CI -- not part of info.rkt and how this would be built if/when used
someday as installed from a catalog pkg server.