Reorganize db initialization
At the module top level, just set a variable to a promise created by
delay/thread, in which the db initialization takes place. The "dbc"
variable is now, instead of a value, a function which forces the
promise.

The idea is to make merely loading the module, faster. I've used a
similar approach for load-setup-xref-collections, which can be slow.