get-point-info: Various changes
Rename some hash keys to clarify which are point-specific or not.

NOTE: If you're also using racket-pdb-mode from the "pdb" branch of
the racket-mode rpeo, you'll need to fetch a new commit there to go
along with this.

Return some new values -- def-sites and doc-sites -- which are
intended to be used for "semantic highlighting".