Add interval-map modules and refine completion candidates
Add new field file-pdb-modules which is an interval-map of module
srclocs.

NOTE: For this to work, a PR needs to be merged in Racket itself:

  https://github.com/racket/racket/pull/4646

A new public API get-submodule-names gives the list of submodule names
for any position. Racket Mode can use this instead of assuming
s-expressions for its "run/enter module at point" command. Likewise
other tools.

Change file-pdb-imports from one set of symbols, to hash from module
path to symbols. In other words, we track import symbols for each
module.

The existing get-completion-candidates public API now requires a
position argument. This is used to return a list of completion
candidates where imports are limited to those valid for the module
around point.

In conjunction with this, when analysis results in errors and no
imports, retain the imports from previous cached analysis if any.
Supports use case of end user editing a file, and wanting completion
candidates even then the file is temporarily in an error state.