Handle nested/complicated prefix-in
Although simple cases of (require (prefix-in)) fully expand
to (#%require (prefix)), for which drracket/check-syntax already gives
two arrows, the general case expands to (#%require (rename)).

To handle that, our updated PR for racket/racket adds a
prefix-sub-range-binders property to the new name, allowing us to
recover the pieces, even for arbitrarily nested prefix-in:

  https://github.com/racket/racket/pull/4649

Here in pdb this commit uses that.