pdb-sub-range-binders => pdb-sub-ranges
Rename this field now that we're storing syntax property values from
import-or-export-prefix-ranges as well as sub-range-binders.

Also simplify the value; an interval-map is overkill when these
rarely contain more than two or three items.

Also store as <pos span> instead of <beg end>, since the same span
applies to both the "use and def ends".