Don't store export prefix ranges in file-pdb-sub-ranges
Previously we were dumping our new stx prop values in
file-pdb-sub-ranges, the same place as sub-range-binders values. This
made some sense when we thought we would use sub-range-binders also
for prefix-{in out}.

But especially now that we're using another stx prop, cleaner to store
the value in a field of the prefixed-export struct, which goes in
file-pdb-exports field. That way, when we encounter a prefixed-export,
we already have the info, w/o needing to go look it up elsewhere.