Store module->exports in the db
An earlier commit shifted away from storing a copy of the symbols in
every analyzed file, which was good.

But it assumed module->exports could be called, later, as quickly as
we can during expansion. That is definitely not the case. When the
module is already declared, as an imported module will be during
expansion, then module->exports is extremely fast. But declaring a
module in a fresh namespace can be quite slow.

In other words, having already paid the cost during expansion, we cash
that in to get the symbols quickly, and cache them.

Note that it's still helpful to resolve the module path, just for a
different reason: It disambiguates relative submodule paths
like (submob "." m).