Move file-add-arrows to its own source file
It felt weird to have this in data-types.rkt, which was mostly about
data definitions. It was in the same file to avoid a mutual require.
The hack to avoid that now is to use a parameter.

Also: Remove use-stx-datum field from syncheck-arrow struct. This
could/should have been in previous commit, where we stopped using it
to detect prefix arrows from check-syntax.