Refine find-uses-in-file
Improve comment.

Gather exports in an intermediate set, to avoid accidental duplicate
calls to find-uses-of-export, which could be exponential; this is
how/when we "fan out" to examine more files, each of which could fan
out, and so on.

Currently that shouldn't happen, due to the nesting of the for
comprehensions and the use of for/or. But if that gets changed
mistakenly in the future, such that the for loops do more work than
necessary, the intermediate set should at least guard against the much
more expensive problem of calling find-export redundantly.

TL;DR: This is some defensive coding for the future.