Use fuller index for finding uses of exports
Previously had a "side index" in the db to quickly find which files
nominally import an export. Then for each file we could get the `file`
struct, and use its arrows.

Although this worked, it could become slow when many and/or large
files were involved, due to the time to unzip/deserialize/read each
file struct. For larger files like class-internal.rkt, this could take
on the order of 1 second, plus result in enough churn to trigger major
GCs fairly frequently.

So instead, now we're storing a fuller index: All the imports, all the
exports, and (to handle re-provides) all the re-exports. A recursive
CTE query can chase down arbitrary chains of re-provides.

This side index gets updated after every analysis. To a first
approximation, all the import-arrows go into the imports table, and
all the pdb-exports items go into the exports table.

In this new approach, def->uses/same-name still follows the chain to
the definition site via the file arrows, analyzing files for the first
time as necessary. And arrows within that defining file are used to
find use sites within the file. Then for each such site, the side
index is used to find uses external to the file, i.e. "fan back out".

[Is it a little weird to use one approach for traversing to the
definition, and such a different one to fan back out? Yes. Does this
raise the question, why not go all the way back the original design
where _everything_ was just rows in db tables? Maybe we could. But I
abandoned that because it was taking too long to insert so many rows
in the db. (Whereas even with this moderately more insert-age, the
time seems OK.) Also, it was a relief to work through all the various
edge cases -- re-provides, anonymous re-provides, sub-range-binders,
prefixed imports and exports, etc. etc. etc. -- using my own system of
arrow structs, as opposed to either raw syncheck arrow values or their
equivalent db columns. So although anything is possible in the future,
this is where I'm at for now.]