Don't use racket/phase+space; re-sync with traversals.rkt
At one point I'd updated the "level" and "level-of-enclosing-module"
vars to be phase+space, expecting that would be necessary for spaces.

However traversals.rkt never got updated that way, so apparently this
is an unnecessary complication. And I want to keep things more similar
to traversals.rkt, with a view toward someday merging most/all of this
to there.

All this was motivated by me fixing an issue with the module* form
where the lang is #f, and this code had been shifting the phase by a
hard-coded -1. Instead it should be (- level), i.e. shift back to
phase 0, like traversals.rkt does. That fix is in this commit, too.