Fix resolved_module_path_exports table constraint
Although I haven't worked with this code in many months, the correct
constraint seems like it must be a unique primary key on the `rmp`
field -- i.e. this is a lookup table with one expected result.

In fact I saw a `match` failure due to a query on this table
unexpectedly returning 2 rows instead of 1.

I'm not sure what I was thinking with the old constraint... was it
just a dumb mistake? I think so.